[aptos-cli] Split out some dependencies to aptos-cli-base
This allows others to build on top of parsers for the config
as well as allows us to reuse code across multiple CLIs.

Additionally, I've cut down the number of dependencies, so
it hopefully can build faster if we split the tool into
multiple pieces.
