+++ crates/aptos-cli-base/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "aptos-cli-base"
+version = "0.1.0"
+authors = ["Aptos Labs <opensource@aptoslabs.com>"]
+description = "CLI common constructs to remove dependencies"
+repository = "https://github.com/aptos-labs/aptos-core"
+homepage = "https://aptoslabs.com"
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0.58"
+bcs = "0.1.3"
+clap = "3.2.11"
+dirs = "4.0.0"
+itertools = "0.10.3"
+serde = "1.0.140"
+serde_json = "1.0.82"
+serde_yaml = "0.8.26"
+thiserror = "1.0.31"
+
+aptos-crypto = { path = "../aptos-crypto" }
+aptos-types = { path = "../../types" }
