+++ crates/aptos-cli-base/src/lib.rs
@@ -1,0 +1,9 @@
+// Copyright (c) Aptos
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod config;
+pub mod file;
+pub mod parse;
+pub mod prompts;
+pub mod types;
+pub mod utils;
