+++ crates/aptos-cli-base/src/types.rs
@@ -1,0 +1,119 @@
+// Copyright (c) Aptos
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::Serialize;
+use thiserror::Error;
+
+/// A common result to be returned to users
+pub type CliResult = Result<String, String>;
+
+/// A common result to remove need for typing `Result<T, CliError>`
+pub type CliTypedResult<T> = Result<T, CliError>;
+
+/// CLI Errors for reporting through telemetry and outputs
+#[derive(Debug, Error)]
+pub enum CliError {
+    #[error("Aborted command")]
+    AbortedError,
+    #[error("API error: {0}")]
+    ApiError(String),
+    #[error("Error (de)serializing '{0}': {1}")]
+    BCS(&'static str, #[source] bcs::Error),
+    #[error("Invalid arguments: {0}")]
+    CommandArgumentError(String),
+    #[error("Unable to load config: {0} {1}")]
+    ConfigLoadError(String, String),
+    #[error("Unable to find config {0}, have you run `aptos init`?")]
+    ConfigNotFoundError(String),
+    #[error("Error accessing '{0}': {1}")]
+    IO(String, #[source] std::io::Error),
+    #[error("Move compilation failed: {0}")]
+    MoveCompilationError(String),
+    #[error("Move unit tests failed")]
+    MoveTestError,
+    #[error("Move Prover failed")]
+    MoveProverError,
+    #[error("Unable to parse '{0}': error: {1}")]
+    UnableToParse(&'static str, String),
+    #[error("Unable to read file '{0}', error: {1}")]
+    UnableToReadFile(String, String),
+    #[error("Unexpected error: {0}")]
+    UnexpectedError(String),
+}
+
+impl CliError {
+    pub fn unexpected<Err: std::fmt::Display>(err: Err) -> CliError {
+        CliError::UnexpectedError(err.to_string())
+    }
+
+    pub fn to_str(&self) -> &'static str {
+        match self {
+            CliError::AbortedError => "AbortedError",
+            CliError::ApiError(_) => "ApiError",
+            CliError::BCS(_, _) => "BCS",
+            CliError::CommandArgumentError(_) => "CommandArgumentError",
+            CliError::ConfigLoadError(_, _) => "ConfigLoadError",
+            CliError::ConfigNotFoundError(_) => "ConfigNotFoundError",
+            CliError::IO(_, _) => "IO",
+            CliError::MoveCompilationError(_) => "MoveCompilationError",
+            CliError::MoveTestError => "MoveTestError",
+            CliError::MoveProverError => "MoveProverError",
+            CliError::UnableToParse(_, _) => "UnableToParse",
+            CliError::UnableToReadFile(_, _) => "UnableToReadFile",
+            CliError::UnexpectedError(_) => "UnexpectedError",
+        }
+    }
+}
+
+impl From<anyhow::Error> for CliError {
+    fn from(err: anyhow::Error) -> Self {
+        CliError::unexpected(err)
+    }
+}
+
+impl From<serde_yaml::Error> for CliError {
+    fn from(err: serde_yaml::Error) -> Self {
+        CliError::unexpected(err)
+    }
+}
+
+impl From<std::string::FromUtf8Error> for CliError {
+    fn from(err: std::string::FromUtf8Error) -> Self {
+        CliError::unexpected(err)
+    }
+}
+
+impl From<bcs::Error> for CliError {
+    fn from(err: bcs::Error) -> Self {
+        CliError::unexpected(err)
+    }
+}
+
+/// A result wrapper for displaying either a correct execution result or an error.
+///
+/// The purpose of this is to have a pretty easy to recognize JSON output format e.g.
+///
+/// {
+///   "Result":{
+///     "encoded":{ ... }
+///   }
+/// }
+///
+/// {
+///   "Error":"Failed to run command"
+/// }
+///
+#[derive(Debug, Serialize)]
+pub enum ResultWrapper<T> {
+    Result(T),
+    Error(String),
+}
+
+impl<T> From<CliTypedResult<T>> for ResultWrapper<T> {
+    fn from(result: CliTypedResult<T>) -> Self {
+        match result {
+            Ok(inner) => ResultWrapper::Result(inner),
+            Err(inner) => ResultWrapper::Error(inner.to_string()),
+        }
+    }
+}
