+++ crates/aptos/Cargo.toml
@@ -11,15 +11,12 @@ edition = "2018"
 build = "build.rs"
 
 [dependencies]
-anyhow = "1.0.57"
 async-trait = "0.1.53"
 base64 = "0.13.0"
 bcs = "0.1.3"
 clap = "3.2.11"
 clap_complete = "3.2.3"
-dirs = "4.0.0"
 hex = "0.4.3"
-itertools = "0.10.3"
 parse_duration = "2.1.1"
 rand = "0.7.3"
 reqwest = { version = "0.11.10", features = ["blocking", "json"] }
@@ -27,13 +24,10 @@ serde = "1.0.137"
 serde_json = "1.0.81"
 serde_yaml = "0.8.24"
 shadow-rs = "0.11.0"
-tempfile = "3.3.0"
-thiserror = "1.0.31"
-tokio = { version = "1.18.2", features = ["full"] }
-tokio-util = { version = "0.7.2", features = ["compat"] }
+tokio =  "1.18.2"
 toml = "0.5.9"
-uuid = { version = "1.0.0", features = ["v4", "serde"] }
 
+aptos-cli-base = { path = "../aptos-cli-base" }
 aptos-config = { path = "../../config" }
 aptos-crypto = { path = "../aptos-crypto", features = [] }
 aptos-faucet = { path = "../aptos-faucet" }
@@ -52,14 +46,8 @@ aptos-transaction-builder = { path = "../../sdk/transaction-builder" }
 aptos-types = { path = "../../types" }
 aptos-vm = { path = "../../aptos-move/aptos-vm" }
 
-aptosdb = { path = "../../storage/aptosdb" }
 cached-framework-packages =  { path = "../../aptos-move/framework/cached-packages" }
-executor = { path = "../../execution/executor" }
-framework = { path = '../../aptos-move/framework' }
 move-deps = { path = "../../aptos-move/move-deps", features = ["address32", "testing", "table-extension"] }
-short-hex-str = { path = "../short-hex-str" }
-storage-interface = { path = "../../storage/storage-interface" }
-vm-genesis = { path = "../../aptos-move/vm-genesis" }
 
 [build-dependencies]
 shadow-rs = "0.11.0"
