+++ crates/aptos/src/account/mod.rs
@@ -1,7 +1,8 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::common::types::{CliCommand, CliResult};
+use crate::common::types::CliCommand;
+use aptos_cli_base::types::CliResult;
 use clap::Subcommand;
 
 pub mod create;
