+++ crates/aptos/src/config/mod.rs
@@ -1,22 +1,15 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::common::types::{CliCommand, CliError, CliResult, CliTypedResult, CONFIG_FOLDER};
-use crate::common::utils::{
-    create_dir_if_not_exist, current_dir, read_from_file, write_to_user_only_file,
-};
-use crate::genesis::git::{from_yaml, to_yaml};
+use crate::common::types::CliCommand;
 use crate::Tool;
+use aptos_cli_base::config::{ConfigType, GlobalConfig};
+use aptos_cli_base::types::{CliError, CliResult, CliTypedResult};
 use async_trait::async_trait;
-use clap::ArgEnum;
 use clap::CommandFactory;
 use clap::Parser;
 use clap_complete::{generate, Shell};
-use serde::Deserialize;
-use serde::Serialize;
-use std::fmt::Formatter;
 use std::path::PathBuf;
-use std::str::FromStr;
 
 /// Tool for configuration of the CLI tool
 ///
@@ -113,100 +106,5 @@ impl CliCommand<GlobalConfig> for ShowGlobalConfig {
     async fn execute(self) -> CliTypedResult<GlobalConfig> {
         // Load the global config
         GlobalConfig::load()
-    }
-}
-
-const GLOBAL_CONFIG_FILE: &str = "global_config.yaml";
-
-/// A global configuration for global settings related to a user
-#[derive(Serialize, Deserialize, Debug, Default)]
-pub struct GlobalConfig {
-    /// Whether to be using Global or Workspace mode
-    pub config_type: ConfigType,
-}
-
-impl GlobalConfig {
-    pub fn load() -> CliTypedResult<Self> {
-        let path = global_folder()?.join(GLOBAL_CONFIG_FILE);
-        if path.exists() {
-            from_yaml(&String::from_utf8(read_from_file(path.as_path())?)?)
-        } else {
-            // If we don't have a config, let's load the default
-            Ok(GlobalConfig::default())
-        }
-    }
-
-    /// Get the config location based on the type
-    pub fn get_config_location(&self) -> CliTypedResult<PathBuf> {
-        match self.config_type {
-            ConfigType::Global => global_folder(),
-            ConfigType::Workspace => Ok(current_dir()?.join(CONFIG_FOLDER)),
-        }
-    }
-
-    fn save(&self) -> CliTypedResult<()> {
-        let global_folder = global_folder()?;
-        create_dir_if_not_exist(global_folder.as_path())?;
-
-        write_to_user_only_file(
-            global_folder.join(GLOBAL_CONFIG_FILE).as_path(),
-            "Global Config",
-            &to_yaml(&self)?.into_bytes(),
-        )
-    }
-}
-
-fn global_folder() -> CliTypedResult<PathBuf> {
-    if let Some(dir) = dirs::home_dir() {
-        Ok(dir.join(CONFIG_FOLDER))
-    } else {
-        Err(CliError::UnexpectedError(
-            "Unable to retrieve home directory".to_string(),
-        ))
-    }
-}
-
-const GLOBAL: &str = "global";
-const WORKSPACE: &str = "workspace";
-
-/// A configuration for where to place and use the config
-///
-/// Workspace allows for multiple configs based on location, where
-/// Global allows for one config for every part of the code
-#[derive(Debug, Copy, Clone, Serialize, Deserialize, ArgEnum)]
-pub enum ConfigType {
-    /// Per system user configuration put in `<HOME>/.aptos`
-    Global,
-    /// Per directory configuration put in `<CURRENT_DIR>/.aptos`
-    Workspace,
-}
-
-impl Default for ConfigType {
-    fn default() -> Self {
-        // TODO: When we version up, we can change this to global
-        Self::Workspace
-    }
-}
-
-impl std::fmt::Display for ConfigType {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.write_str(match self {
-            ConfigType::Global => GLOBAL,
-            ConfigType::Workspace => WORKSPACE,
-        })
-    }
-}
-
-impl FromStr for ConfigType {
-    type Err = CliError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().trim() {
-            GLOBAL => Ok(Self::Global),
-            WORKSPACE => Ok(Self::Workspace),
-            _ => Err(CliError::CommandArgumentError(
-                "Invalid config type, must be one of [global, workspace]".to_string(),
-            )),
-        }
     }
 }
